@@ -1,19 +1,19 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
+apply plugin: 'kotlin-parcelize'
 apply plugin: 'androidx.navigation.safeargs.kotlin'
 apply plugin: 'kotlin-kapt'
 
 android {
-    compileSdkVersion 29
+    compileSdkVersion 30
     buildToolsVersion "29.0.3"
 
     defaultConfig {
         applicationId "com.github.newscalculator"
         minSdkVersion 21
-        targetSdkVersion 29
-        versionCode 4
-        versionName "1.0.2 beta"
+        targetSdkVersion 30
+        versionCode 5
+        versionName "1.0.3 beta"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -36,11 +36,7 @@ android {
         jvmTarget = '1.8'
     }
 
-    androidExtensions {
-        experimental = true
-    }
     buildFeatures {
-        dataBinding = true
         viewBinding = true
     }
 
@@ -50,16 +46,29 @@ dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.core:core-ktx:1.3.2'
     implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'com.google.android.material:material:1.2.1'
+    implementation 'com.google.android.material:material:1.3.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
     testImplementation 'junit:junit:4.13.1'
     implementation "androidx.cardview:cardview:1.0.0"
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+
+    //ViewModel
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.2.0"
     implementation "androidx.fragment:fragment-ktx:1.2.5"
+
+    //DelegatesAdapter
     implementation 'com.hannesdorfmann:adapterdelegates4:4.3.0'
+
+    //LiveData
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.2.0"
-    implementation 'com.google.android.material:material:1.2.1'
-    implementation "androidx.navigation:navigation-fragment-ktx:2.3.2"
+
+    //Navigation
+    implementation "androidx.navigation:navigation-fragment-ktx:2.3.3"
+
+    //Moshi
+    def moshiVersion = '1.9.3'
+    implementation "com.squareup.moshi:moshi:$moshiVersion"
+    implementation "com.squareup.moshi:moshi-kotlin:$moshiVersion"
+    kapt "com.squareup.moshi:moshi-kotlin-codegen:$moshiVersion"
 }
